Add score tally and declaration of overall winner
scoreUpdate function created to increment the score tally
of the round winner and display the updated score on the
webpage. Function also checks if there is an overall winner
(i.e., score equal to WINNING_SCORE constant) and announces
the winner. Once an overall winner is determined, the game
is stopped by removal of the event listeners.